Tests: Moved common code to separate module, in order to reduce boilerplate amount
